(chore) - set scripts for production
- ensure transpling to ES6 with babel
- add scripts for creating models
- set production scripts
[Chore #165484968]
