+++ Procfile
@@ -1,1 +1,1 @@
-web: npm run migrate && npm run production
+web: npm run production
