+++ package.json
@@ -9,11 +9,12 @@
     "start": "set NODE_ENV=production && npm run migrate && node server/bin/dev",
     "clean": "rm -rf dist",
     "build": "npm run clean && mkdir dist && babel server -s -d dist",
-    "production": "npm run build && node server/bin/production",
-    "migrate": "babel-node db createAllTables",
-    "insert": "babel-node testdata insertData",
-    "rollback": "babel-node db dropAllTables",
-    "mockdata": "npm run rollback && npm run migrate && npm run insert",
+    "production": "npm run migrate && node server/bin/production",
+    "migrate": "npm run build && node dist/sqldatabase/models createAllTables",
+    "create": "node dist/sqldatabase/models createAllTables",
+    "insert": "node dist/sqldatabase/testdata insertData",
+    "rollback": "node dist/sqldatabase/models dropAllTables",
+    "mockdata": "npm run build && npm run rollback && npm run create && npm run insert",
     "generate-lcov": "nyc report --reporter=text-lcov > lcov.info",
     "coveralls-coverage": "coveralls < lcov.info",
     "codeclimate-coverage": "codeclimate-test-reporter < lcov.info",
